refactor: consolidate compaction strategy types and remove obsolete strategy_type module
