+++ crates/forge_domain/src/compaction/mod.rs
@@ -9,7 +9,6 @@ mod context_compactor;
 pub mod impact;
 mod sliding_window;
 mod strategy;
-mod strategy_type;
 mod summarization;
 
 pub use context_compactor::ContextCompactor;
