+++ crates/forge_domain/src/compaction/strategy.rs
@@ -1,4 +1,9 @@
+use anyhow::Result;
 use async_trait::async_trait;
+
+use super::sliding_window::SlidingWindowStrategy;
+use super::summarization::SummarizationStrategy;
+use crate::{Context, Services};
 
 /// User-defined compaction strategy for managing context size
 /// Each implementation provides a different approach to context
@@ -19,4 +24,37 @@ pub trait CompactionStrategy: Send + Sync {
         compact: &crate::Compact,
         context: crate::Context,
     ) -> anyhow::Result<crate::Context>;
+}
+
+
+/// The available compaction strategies
+pub enum StrategyType<S> {
+    Summarization(SummarizationStrategy<S>),
+    SlidingWindow(SlidingWindowStrategy),
+}
+
+impl<S: Services> StrategyType<S> {
+    /// Get the ID of the strategy
+    pub fn id(&self) -> &'static str {
+        match self {
+            StrategyType::Summarization(s) => s.id(),
+            StrategyType::SlidingWindow(s) => s.id(),
+        }
+    }
+
+    /// Check if this strategy is applicable to the given context
+    pub fn is_applicable(&self, compact: &crate::Compact, context: Context) -> bool {
+        match self {
+            StrategyType::Summarization(s) => s.is_applicable(compact, context),
+            StrategyType::SlidingWindow(s) => s.is_applicable(compact, context),
+        }
+    }
+
+    /// Apply the compaction strategy
+    pub async fn compact(&self, compact: &crate::Compact, context: Context) -> Result<Context> {
+        match self {
+            StrategyType::Summarization(s) => s.compact(compact, context).await,
+            StrategyType::SlidingWindow(s) => s.compact(compact, context).await,
+        }
+    }
 }
